requirements.txt
requests
python-dotenv
jinja2
fastapi[standard]
python-multipart
pytest
pydantic
